backend/requirements.txt
Flask
flask-cors
pandas
openpyxl
Pillow
pytesseract
tabula-py
